Завершение DI[Добавление фабрик]

+++ HomeTask/battleships/ContainerModules.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Ninject;
+using Ninject.Parameters;
 
 namespace battleships
 {
@@ -14,7 +15,7 @@ namespace battleships
 	        Bind<IGameVisualizer>().To<GameVisualizer>();
             Bind<IProcessMonitor>().To<ProcessMonitor>()
                  .WithConstructorArgument(TimeSpan.FromSeconds(settings.TimeLimitSeconds * settings.GamesCount))
-                 .WithConstructorArgument(settings.MemoryLimit);
+                 .WithConstructorArgument((long)settings.MemoryLimit);
 
             Bind<AiTester>().ToSelf().WithConstructorArgument(settings);
         }
